feat(ReactObjectForm): handling for properties of type 'number'
Properties of type number result in using the NumberValueInput-component internally. This renders a

empty input for Number.NaN and parses the input with parseFloat(). Only valid numbers trigger a

'onChange' event and update the state of the input element. A empty input returns 'Number.NaN'.

Numbers are not converted to string properties anymore!